+++ src/main/java/data/DukeException.java
@@ -1,3 +1,5 @@
+package data;
+
 public class DukeException extends Exception {
     public DukeException(String errMsg) {
         super(errMsg);
