+++ src/main/java/task/Deadline.java
@@ -1,7 +1,9 @@
+package task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
     protected LocalDate by;
     public Deadline(String task, LocalDate by) {
         super(task.trim());
@@ -35,6 +37,6 @@ public class Deadline extends Task{
     }
     @Override
     public String toString() {
-        return "["+Type.D+"]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) + ")";
+        return "["+ Type.D+"]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) + ")";
     }
 }
