+++ src/main/java/task/TaskList.java
@@ -1,7 +1,9 @@
+package task;
+
 import java.util.ArrayList;
 
 public class TaskList {
-    protected ArrayList<Task> tasks;
+    public ArrayList<Task> tasks;
 
     public TaskList() {
     }
