+++ src/main/java/task/Type.java
@@ -1,3 +1,5 @@
+package task;
+
 public enum Type {
     T, D, E
 }
