+++ text-ui-test/EXPECTED.TXT
@@ -1,29 +1,25 @@
- ____        _
-|  _ \ _   _| | _____
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
+    (/ /^ / - / ^/ /)
     ____________________________________________________________
     Hello! I'm Duke by A0221330A.
     What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task:
+    Got it. I've added this task: 
         [T][ ] read book
     Now you have 1 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task:
-        [D][ ] return book (by: June 6th)
+    Got it. I've added this task: 
+        [D][ ] return book (by: Dec 20 2022)
     Now you have 2 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task:
-        [E][ ] project meeting (at: Aug 6th 2-4pm)
+    Got it. I've added this task: 
+        [E][ ] project meeting (at: Dec 21 2022)
     Now you have 3 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task:
+    Got it. I've added this task: 
         [T][ ] join sports club
     Now you have 4 tasks in the list.
     ____________________________________________________________
@@ -36,27 +32,17 @@
         [T][✓] join sports club
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task:
+    Got it. I've added this task: 
         [T][ ] borrow book
     Now you have 5 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
     Here are the tasks in your list:
     1. [T][✓] read book
-    2. [D][ ] return book (by: June 6th)
-    3. [E][ ] project meeting (at: Aug 6th 2-4pm)
+    2. [D][ ] return book (by: Dec 20 2022)
+    3. [E][ ] project meeting (at: Dec 21 2022)
     4. [T][✓] join sports club
     5. [T][ ] borrow book
-    ____________________________________________________________
-    ____________________________________________________________
-    Got it. I've added this task:
-        [D][ ] return book (by: Sunday)
-    Now you have 6 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-    Got it. I've added this task:
-        [E][ ] project meeting (at: Mon 2-4 pm)
-    Now you have 7 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
     Bye. Hope to see you again soon!
